Rename Login.vue become to AccountButton.vue and Button.vue become to UserButton.vue
